+++ tests/GH27Spec.hs
@@ -22,6 +22,11 @@ We isolate this test case from the rest of "DerivingViaSpec" because
 we have to disable @-Wunused-foralls@ due to
 https://gitlab.haskell.org/ghc/ghc/issues/13512, and we would prefer not to
 contaminate the rest of "DerivingViaSpec" with this hack.
+
+Unfortunately, this test case cannot be made to work yet due to
+https://github.com/haskell-compat/deriving-compat/issues/34, so we leave it
+here commented out. If that issue were to be fixed, we should be able to bring
+this test case back.
 -}
 module GH27Spec where
 
@@ -30,6 +35,7 @@ import Prelude.Compat
 
 import Test.Hspec
 
+{-
 #if MIN_VERSION_template_haskell(2,12,0)
 import Data.Deriving.Via
 import Data.Functor.Const
@@ -37,6 +43,7 @@ import Data.Functor.Const
 newtype Age = MkAge Int
 $(deriveVia [t| forall a. Show Age `Via` Const Int a |])
 #endif
+-}
 
 main :: IO ()
 main = hspec spec